feat(createuseradmin.ts): add function to create a new user with predefined username and password
The `createNewUser` function is added to the `createuseradmin.ts` file. This function is responsible for creating a new user with a predefined username and password. The username is set to "sinacofiadmin" and the password is set to "1Q2w3e4R". This function uses the `PrismaClient` to interact with the database.